+++ freecad/Ratchet/Misc/__init__.py
@@ -1,0 +1,7 @@
+# SPDX-License-Identifier: LGPL-3.0-or-later
+# SPDX-FileNotice: Part of the Ratchet addon.
+
+from .Resources import Paths , asIcon
+from .Document import DocumentSwitch
+from .Logging import report
+from .Version import Version