+++ freecad/Ratchet/PySide/QtWidgets.py
@@ -1,0 +1,10 @@
+# SPDX-License-Identifier: LGPL-3.0-or-later
+# SPDX-FileNotice: Part of the Ratchet addon.
+
+from typing import TYPE_CHECKING
+
+
+if TYPE_CHECKING:
+    from PySide6.QtWidgets import *
+else:
+    from PySide.QtWidgets import *